refactor: update shims to support absolute imports
