+++ api/client/gro_client/__init__.py
@@ -1,2 +1,2 @@
-from groapiclient.client import GroClient
+from groapiclient import GroClient
 from groapiclient.client import APIError
