+++ api/client/lib/__init__.py
@@ -1,0 +1,1 @@
+from groapiclient.lib import *
