+++ groapiclient/__init__.py
@@ -1,2 +1,2 @@
-from groapiclient import GroClient as Client
+from groapiclient.client import GroClient
 from groapiclient import lib
